Breaking down codes into pages done
